+++ src/types.d.ts
@@ -2,3 +2,8 @@ declare module "*.sql" {
   const text: string;
   export default text;
 }
+
+declare module "*.md" {
+  const text: string;
+  export default text;
+}